backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
scikit-learn
pandas
numpy
transformers
torch
pdfplumber
pytesseract
pillow
python-dotenv
aiohttp
beautifulsoup4
opencv-python
python-Levenshtein
sqlalchemy
pymysql
celery
redis
slowapi
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
pytest
pytest-asyncio
alembic
datasets
accelerate
